backend/app/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
PyJWT==2.3.0
PyJWT==2.8.0
